gmail-slack-integration/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
slack-sdk==3.23.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
apscheduler==3.10.4
httpx==0.25.2